+++ src/olmo_core/train/config.py
@@ -101,6 +101,7 @@ class TrainerConfig(Config):
             collator=collator,
             checkpointer=checkpointer,
             train_sequence_length=dataset.sequence_length,
+            max_train_sequence_length=dataset.max_target_sequence_length,
             autocast_precision=None if autocast_precision is None else autocast_precision.as_pt(),
             work_dir=Path(work_dir),
             device=torch.device(device) if device is not None else get_default_device(),